% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_license.R
\name{build_license}
\alias{build_license}
\title{Build License}
\usage{
build_license(license, ...)
}
\arguments{
\item{license}{Character string specifying the license name}

\item{...}{Additional arguments passed to underlying usethis function}
}
\value{
Invisible TRUE if successful
}
\description{
A unified interface to usethis license functions. Uses match_license() to map
license input to a supported license, then calls the appropriate usethis function.
}
