% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_package.R
\name{build_roxygen}
\alias{build_roxygen}
\title{Generate roxygen documentation for datasets}
\usage{
build_roxygen(type = "dataset", base_path = NULL, verbose = TRUE)
}
\arguments{
\item{type}{Type of documentation to generate (default: "dataset")}

\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Logical indicating success
}
\description{
Creates R documentation files for each dataset using the data dictionary.
}
