% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_site}
\alias{build_site}
\title{Build website with pkgdown}
\usage{
build_site(base_path = NULL, verbose = TRUE, preview = TRUE, install = TRUE)
}
\arguments{
\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages (default: TRUE)}

\item{preview}{Whether to preview the site (default: TRUE)}

\item{install}{Whether to install the package (default: TRUE)}
}
\value{
Logical indicating success
}
\description{
Configures pkgdown for building the package's website.
}
