% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checklist_summary}
\alias{checklist_summary}
\title{Get summary string for validation results}
\usage{
checklist_summary(validation)
}
\arguments{
\item{validation}{Validation results}
}
\value{
Character string with summary
}
\description{
Get summary string for validation results
}
