% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{create_citation}
\alias{create_citation}
\title{Create CITATION file using cffr}
\usage{
create_citation(base_path = NULL, verbose = TRUE, validate = FALSE)
}
\arguments{
\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\description{
Creates a CITATION file from DESCRIPTION using the cffr package.
}
