% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_package.R
\name{.create_directories}
\alias{.create_directories}
\title{Create project directories}
\usage{
.create_directories(
  dirs = c("data_raw", "data", "inst/extdata"),
  base_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dirs}{Character vector of directories to create}

\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Character vector of created directories
}
\description{
Creates the necessary directory structure for a fairenough project.
}
