% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{export_rda}
\alias{export_rda}
\title{Export data as RDA file}
\usage{
export_rda(data, name, base_path = NULL, overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{Data frame to export}

\item{name}{Name for the data object (without extension)}

\item{base_path}{Base path for the project}

\item{overwrite}{Whether to overwrite existing file (default: TRUE)}

\item{verbose}{Whether to show messages (default: TRUE)}
}
\value{
Path to the created file
}
\description{
Saves a data frame as an .rda file in the data directory.
}
