% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_supported_files}
\alias{filter_supported_files}
\title{Filter files by supported extensions}
\usage{
filter_supported_files(file_paths)
}
\arguments{
\item{file_paths}{Character vector of file paths}
}
\value{
Character vector of files with supported extensions
}
\description{
Filter files by supported extensions
}
