% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R, R/build_legacy.R
\name{generate_data_documentation}
\alias{generate_data_documentation}
\title{Generate roxygen documentation for datasets}
\usage{
generate_data_documentation(base_path = NULL, verbose = TRUE)

generate_data_documentation(base_path = NULL, verbose = TRUE)
}
\arguments{
\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Logical indicating success

Logical indicating success
}
\description{
Creates R documentation files for each dataset using the data dictionary.

Creates R documentation files for each dataset using the data dictionary.
}
