% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R, R/build_legacy.R
\name{generate_readme}
\alias{generate_readme}
\title{Generate README from template}
\usage{
generate_readme(base_path = NULL, verbose = TRUE)

generate_readme(base_path = NULL, verbose = TRUE)
}
\arguments{
\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Path to generated README

Path to generated README
}
\description{
Creates README.md from a template, incorporating package metadata.

Creates README.md from a template, incorporating package metadata.
}
