% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_license.R
\name{get_available_licenses}
\alias{get_available_licenses}
\title{Get Available License Names}
\usage{
get_available_licenses()
}
\value{
Character vector of supported license names
}
\description{
Returns a character vector of all supported license names.
}
\examples{
get_available_licenses()
}
