% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata_from_desc}
\alias{get_metadata_from_desc}
\title{Get metadata from DESCRIPTION file}
\usage{
get_metadata_from_desc(base_path = NULL)
}
\arguments{
\item{base_path}{Base path for the project}
}
\value{
List containing metadata
}
\description{
Reads DESCRIPTION file and reconstructs metadata structure
}
