% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_supported_extensions}
\alias{get_supported_extensions}
\title{Get supported file extensions}
\usage{
get_supported_extensions()
}
\value{
Character vector of supported file extensions (without dots)
}
\description{
Get supported file extensions
}
