% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data_raw.R
\name{init_data_raw}
\alias{init_data_raw}
\title{Initialize data_raw directory}
\usage{
init_data_raw(data_raw_dir = "data_raw", gitignore = TRUE, base_path = NULL)
}
\arguments{
\item{data_raw_dir}{Name of the raw data directory}

\item{gitignore}{Whether to add directory to .gitignore}

\item{base_path}{Base path for the project}
}
\description{
Initialize data_raw directory
}
