% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_export.R
\name{init_export}
\alias{init_export}
\title{Export data file to package data}
\usage{
init_export(
  data_file,
  overwrite_rda = TRUE,
  auto_clean = TRUE,
  base_path = NULL
)
}
\arguments{
\item{data_file}{Path to a CSV/Excel file}

\item{overwrite_rda}{Whether to overwrite existing .rda files}

\item{auto_clean}{Whether to auto-clean data}

\item{base_path}{Base path for the project}
}
\description{
Export data file to package data
}
