% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_build_completed}
\alias{is_build_completed}
\title{Check if build step is completed}
\usage{
is_build_completed(base_path = NULL)
}
\arguments{
\item{base_path}{Base path for the project}
}
\value{
Logical indicating if build is completed
}
\description{
Check if build step is completed
}
