% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_dictionary_generated}
\alias{is_dictionary_generated}
\title{Check if dictionary generation is completed}
\usage{
is_dictionary_generated(base_path = NULL)
}
\arguments{
\item{base_path}{Base path for the project}
}
\value{
Logical indicating if dictionary generation is completed
}
\description{
Check if dictionary generation is completed
}
