% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_metadata_collected}
\alias{is_metadata_collected}
\title{Check if metadata collection is completed}
\usage{
is_metadata_collected(base_path = NULL)
}
\arguments{
\item{base_path}{Base path for the project}
}
\value{
Logical indicating if metadata collection is completed
}
\description{
Check if metadata collection is completed
}
