% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_pkg}
\alias{is_pkg}
\title{Check if current directory is a package directory}
\usage{
is_pkg()
}
\value{
Logical indicating if current directory contains DESCRIPTION and NAMESPACE files
}
\description{
Check if current directory is a package directory
}
