% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_processing_completed}
\alias{is_processing_completed}
\title{Check if data processing step is completed}
\usage{
is_processing_completed(base_path = NULL)
}
\arguments{
\item{base_path}{Base path for the project}
}
\value{
Logical indicating if processing is completed
}
\description{
Check if data processing step is completed
}
