% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_setup_completed}
\alias{is_setup_completed}
\title{Check if setup step is completed}
\usage{
is_setup_completed(base_path = NULL)
}
\arguments{
\item{base_path}{Base path for the project}
}
\value{
Logical indicating if setup is completed
}
\description{
Check if setup step is completed
}
