% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_supported_file_type}
\alias{is_supported_file_type}
\title{Check if file type is supported}
\usage{
is_supported_file_type(file_path)
}
\arguments{
\item{file_path}{Path to file}
}
\value{
Logical indicating if file type is supported
}
\description{
Check if file type is supported
}
