% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_license.R
\docType{data}
\name{LICENSE_CONFIG}
\alias{LICENSE_CONFIG}
\title{License Configuration}
\format{
A list containing license metadata
}
\usage{
LICENSE_CONFIG
}
\description{
Centralized configuration for all supported licenses.
Makes it easy to add/remove licenses in one place.
}
\keyword{datasets}
