% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_object}
\alias{load_object}
\title{Utility functions for fairenough
Load object from file}
\usage{
load_object(file)
}
\arguments{
\item{file}{Path to the R data file}
}
\value{
The loaded object
}
\description{
Utility functions for fairenough
Load object from file
}
