% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{move_data_files}
\alias{move_data_files}
\title{Move data files to raw data directory}
\usage{
move_data_files(raw_dir = "data_raw", base_path = NULL, verbose = TRUE)
}
\arguments{
\item{raw_dir}{Name of the raw data directory}

\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Character vector of moved files
}
\description{
Moves CSV and Excel files from the project root to the data_raw directory.
}
