% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{prepare_cran}
\alias{prepare_cran}
\title{Prepare package for CRAN}
\usage{
prepare_cran(base_path = NULL, verbose = TRUE)
}
\arguments{
\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Logical indicating readiness
}
\description{
Prepares package for CRAN submission.
}
