% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{prepare_github}
\alias{prepare_github}
\title{Prepare package for GitHub}
\usage{
prepare_github(base_path = NULL, verbose = TRUE)
}
\arguments{
\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Logical indicating success
}
\description{
Creates or updates GitHub-specific files.
}
