% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promptit.R
\name{prompt_confirm}
\alias{prompt_confirm}
\title{Yes/No confirmation}
\usage{
prompt_confirm(message, value = NULL, default = TRUE)
}
\arguments{
\item{message}{Question to ask}

\item{value}{Current value - if not NULL, returns it without prompting}

\item{default}{Default choice (TRUE for yes, FALSE for no)}
}
\value{
Logical TRUE or FALSE
}
\description{
Yes/No confirmation
}
