% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promptit.R
\name{prompt_menu}
\alias{prompt_menu}
\title{Menu selection using utils::menu()}
\usage{
prompt_menu(
  choices,
  title = "Select an option",
  value = NULL,
  default = NULL,
  allow_none = FALSE,
  allow_other = FALSE,
  graphics = getOption("menu.graphics", FALSE)
)
}
\arguments{
\item{choices}{Character vector or named vector (names for display, values for return)}

\item{title}{Title to display above menu}

\item{value}{Current value - if not NULL, returns it without prompting}

\item{default}{Default choice (index or value)}

\item{allow_none}{Add "None of the above" option}

\item{allow_other}{Add "Other (specify)" option}

\item{graphics}{Whether to use graphical menu if available}
}
\value{
Selected value or NULL
}
\description{
Menu selection using utils::menu()
}
