% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{run_checks}
\alias{run_checks}
\title{Run package checks}
\usage{
run_checks(base_path = NULL, verbose = TRUE)
}
\arguments{
\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Logical indicating if checks passed
}
\description{
Runs R CMD check on the package.
}
