% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_roxygen.R
\name{setup_roxygen}
\alias{setup_roxygen}
\title{Set up roxygen documentation for all tidy data sets using the dictionary}
\usage{
setup_roxygen(
  title = "Title goes here ...",
  description = "Description here ..."
)
}
\description{
Set up roxygen documentation for all tidy data sets using the dictionary
}
