% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{setup_website}
\alias{setup_website}
\title{Setup package website}
\usage{
setup_website(base_path = NULL, verbose = TRUE, install = TRUE, preview = TRUE)
}
\arguments{
\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages (default: TRUE)}

\item{install}{Whether to install the package before building site (default: TRUE)}

\item{preview}{Whether to preview the site (default: TRUE)}
}
\value{
Logical indicating success
}
\description{
Configures pkgdown for creating a package website.
}
