% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_checklist}
\alias{show_checklist}
\title{Display formatted checklist to user}
\usage{
show_checklist(validation, title, verbose = TRUE)
}
\arguments{
\item{validation}{Validation results from format_checklist_results}

\item{title}{Title for the checklist display}

\item{verbose}{Whether to show detailed output}
}
\description{
Display formatted checklist to user
}
