% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{SUPPORTED_EXTENSIONS}
\alias{SUPPORTED_EXTENSIONS}
\title{Utility functions for fairenough}
\format{
An object of class \code{character} of length 3.
}
\usage{
SUPPORTED_EXTENSIONS
}
\description{
Utility functions for fairenough
}
\keyword{datasets}
