% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_license.R
\docType{data}
\name{SUPPORTED_LICENSES}
\alias{SUPPORTED_LICENSES}
\title{Supported License Names}
\format{
An object of class \code{character} of length 10.
}
\usage{
SUPPORTED_LICENSES
}
\description{
A character vector of canonical license names supported by this wrapper.
}
\keyword{datasets}
