% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_template}
\alias{use_template}
\title{Use a template file with data substitution}
\usage{
use_template(
  template,
  save_as = template,
  data = list(),
  base_path = NULL,
  package = "fairenough",
  open = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{template}{Name of template file in inst/templates}

\item{save_as}{Path to save file relative to base_path}

\item{data}{List of data for substitution}

\item{base_path}{Base path for the project}

\item{package}{Package containing the template}

\item{open}{Whether to open the file after creation}

\item{verbose}{Whether to show messages}
}
\value{
Path to created file
}
\description{
Similar to usethis::use_template but respects base_path.
Reads a template from the package, substitutes data, and writes to target location.
}
