% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_build_completed}
\alias{validate_build_completed}
\title{Validate build completion status}
\usage{
validate_build_completed(base_path = NULL)
}
\arguments{
\item{base_path}{Base path for the project}
}
\value{
Formatted checklist results
}
\description{
Validate build completion status
}
