% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data_frame}
\alias{validate_data_frame}
\title{Validate data frame}
\usage{
validate_data_frame(data, min_rows = 1)
}
\arguments{
\item{data}{Data frame to validate}

\item{min_rows}{Minimum number of rows required (default 1)}
}
\value{
The validated data frame (throws error if invalid)
}
\description{
Validate data frame
}
