% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_dictionary_completed}
\alias{validate_dictionary_completed}
\title{Validate dictionary generation completion status}
\usage{
validate_dictionary_completed(base_path = NULL)
}
\arguments{
\item{base_path}{Base path for the project}
}
\value{
Formatted checklist results
}
\description{
Validate dictionary generation completion status
}
