% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_file_path}
\alias{validate_file_path}
\title{Validate file path}
\usage{
validate_file_path(file_path)
}
\arguments{
\item{file_path}{Path to file}
}
\value{
The validated file path (throws error if invalid)
}
\description{
Validate file path
}
