% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promptit.R
\name{validators}
\alias{validators}
\alias{validate_email}
\alias{validate_package_name}
\alias{validate_orcid}
\alias{validate_date}
\alias{validate_url}
\title{Common validators}
\usage{
validate_email(x)

validate_package_name(x)

validate_orcid(x)

validate_date(x)

validate_url(x)
}
\value{
Function that returns TRUE for valid input, FALSE otherwise
}
\description{
Common validators
}
\section{Functions}{
\itemize{
\item \code{validate_email()}: Email validator

\item \code{validate_package_name()}: Package name validator

\item \code{validate_orcid()}: ORCID validator

\item \code{validate_date()}: Date validator (YYYY or YYYY-MM-DD)

\item \code{validate_url()}: URL validator

}}
