% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{collect_metadata}
\alias{collect_metadata}
\title{Collect package metadata}
\usage{
collect_metadata(interactive = TRUE, base_path = NULL, verbose = TRUE, ...)
}
\arguments{
\item{interactive}{Whether to use interactive prompts}

\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}

\item{...}{Additional arguments passed to doc_metadata}
}
\value{
List containing metadata
}
\description{
Wrapper around doc_metadata for consistent naming.
}
