% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{create_zenodo_json}
\alias{create_zenodo_json}
\title{Create Zenodo metadata file}
\usage{
create_zenodo_json(base_path = NULL, verbose = TRUE)
}
\arguments{
\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Path to created file
}
\description{
Creates a .zenodo.json file with metadata for Zenodo repository.
}
