% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata from DESCRIPTION file}
\usage{
get_metadata(base_path = NULL)
}
\arguments{
\item{base_path}{Base path for the project}
}
\value{
List containing metadata
}
\description{
Reads DESCRIPTION file and reconstructs metadata structure
}
