% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R, R/build_legacy.R
\name{setup_website}
\alias{setup_website}
\title{Setup package website}
\usage{
setup_website(base_path = NULL, verbose = TRUE)

setup_website(base_path = NULL, verbose = TRUE)
}
\arguments{
\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Logical indicating success

Logical indicating success
}
\description{
Configures pkgdown for creating a package website.

Configures pkgdown for creating a package website.
}
