% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_metadata}
\alias{update_metadata}
\title{Update metadata.json with new data}
\usage{
update_metadata(key, data, base_path = NULL, verbose = TRUE)
}
\arguments{
\item{key}{Top-level key in metadata.json (e.g., "datasets", "package")}

\item{data}{Data to add/update for the key}

\item{base_path}{Base path for the project}

\item{verbose}{Whether to show messages}
}
\value{
Updated metadata list (invisibly)
}
\description{
Adds or updates data in metadata.json file. Creates the file if it doesn't exist.
}
